add feature for chat private

+++ src/main/java/teleder/core/dtos/Pagination.java
@@ -7,16 +7,16 @@ public class Pagination {
 
     private long total;
 
-    private int skip;
+    private long skip;
     private int limit;
 
-    public Pagination(long total, int skip, int limit) {
+    public Pagination(long total, long skip, int limit) {
         this.total = total;
         this.skip = skip;
         this.limit = limit;
     }
 
-    public static Pagination create(long total, int skip, int limit) {
+    public static Pagination create(long total, long skip, int limit) {
         return new Pagination(total, skip, limit);
     }
